Rename spawn launcher files (#13743)
